refactor(native): group runtime render services under render subdomain
