+++ src/native/engine/README.md
@@ -105,7 +105,7 @@ monolithic `engine.c` control block:
 - `runtime/engine/engine_tick.*`: tick pipeline execution and phase ordering.
 - `runtime/input/input_contract.*`: click/move intent contract and sanitization.
 - `runtime/move_target_domain.*`: strongly-typed click-target state and steering decisions.
-- `runtime/render_submit.*` + `RendererDrawCommand`: typed scene command
+- `runtime/render/render_submit.*` + `RendererDrawCommand`: typed scene command
   submission seam so domain render intent is backend-agnostic.
 - `runtime/player/player_*.*`, `runtime/terrain/terrain_*.*`, `runtime/world/world_*.*`: bounded
   contexts for player, terrain, and world behaviors.
