+++ src/native/engine/engine.c
@@ -17,7 +17,7 @@
 #include "runtime/move_target_service.h"
 #include "runtime/player/player_runtime_service.h"
 #include "runtime/player/player_gateway_abi.h"
-#include "runtime/render_material.h"
+#include "runtime/render/render_material.h"
 #include "runtime/player/player_motion.h"
 #include "runtime/player/player_motion_host.h"
 #include "runtime/player/player_api.h"
@@ -31,7 +31,7 @@
 #include "runtime/physics_abi.h"
 #include "runtime/input/input_abi.h"
 #include "runtime/resource_domain.h"
-#include "runtime/render_submit.h"
+#include "runtime/render/render_submit.h"
 #include "runtime/terrain/terrain_generation.h"
 #include "runtime/terrain/terrain_abi.h"
 #include "runtime/terrain/terrain_runtime.h"
