+++ src/native/engine/runtime/render/render_submit.h
@@ -1,10 +1,10 @@
 #ifndef BANANA_ENGINE_RUNTIME_RENDER_SUBMIT_H
 #define BANANA_ENGINE_RUNTIME_RENDER_SUBMIT_H
 
-#include "../render/material.h"
-#include "../render/mesh.h"
-#include "../render/renderer.h"
-#include "../world/world.h"
+#include "../../render/material.h"
+#include "../../render/mesh.h"
+#include "../../render/renderer.h"
+#include "../../world/world.h"
 
 #ifdef __cplusplus
 extern "C"
