+++ tests/native/runtime_render_material_test.c
@@ -1,4 +1,4 @@
-#include "runtime/render_material.h"
+#include "runtime/render/render_material.h"
 
 #include <math.h>
 #include <stdio.h>
